WIP: Adds linear casedef constructuction.
* Adds TransformedMethodKey and TransformedMethod to the
* Uses putAttachment to attatch the transformed method to its subtrees
to allow for local variable reconstruction in the transformation
ContinuationsPhase companion
* Constructs a Thicket around the original method rhs. Then collects
the trees until a tree calling shift is encountered into a list of
thickets, each ending in a call to shift. Finally, constructs the
state machine match tree, creating a CaseDef for each thicket, then
appends the final invocation error case.
* Extracts IllegalArgumentException full name to constants
* Extracts resumeBeforeInvoke message to constants